pci: keep the allocation cursors 64-bit at the 4 GiB pool end
The cursors mem, mem_pf and io advanced as 32-bit values: after a
BAR allocation whose end is exactly the pool end 0x100000000 (now
reachable), *base = bar_value + length wrapped to 0.  Every later
allocation then passed the start and end checks and programmed its
BAR at address 0, over the legacy IO range and DRAM.

Widen the three cursors to 64-bit and advance with a 64-bit sum, so
an exhausted pool leaves the cursor at the end.  The address
parameters of pci_enum_next_aligned32 and pci_align_check_up widen
with them; pci_enum_next_aligned32 computes in uint64_t rather than
uintptr_t, which on 32-bit targets would truncate the exhausted
cursor back to 0 and defeat the addr > 0xffffffff rejection.  The
programmed BAR value stays 32-bit.

The 4 GiB pool unit test now also adds a second device with a
preset (previously programmed) BAR: exactly filling the pool must
leave that BAR untouched instead of re-allocating it from a wrapped
cursor.  The mock learns to seed a BAR preset from the bar info,
and the loop variable shadowing in test_pci_commit that it exposed
is fixed (the inner preset loop clobbered the outer node loop
counter, so only the first node was ever committed).

Verified: unit-pci 30/30, unit-pci-4gib green, full nxp_t1024
powerpc build green.

+++ include/pci.h
@@ -89,12 +89,14 @@ typedef  struct {
 } pci_ctrlr_info_t;
 
 struct pci_enum_info {
-    uint32_t mem;
-    /* Exclusive pool ends.  64-bit: a pool may end exactly at 4 GiB
-     * (0x100000000), which a 32-bit field cannot represent. */
+    /* Allocation cursors and exclusive pool ends.  All 64-bit: a pool
+     * may end exactly at 4 GiB (0x100000000), which a 32-bit value
+     * cannot represent, and an exhausted cursor must stay at the pool
+     * end instead of wrapping to 0 and re-allocating over address 0. */
+    uint64_t mem;
     uint64_t mem_limit;
-    uint32_t io;
-    uint32_t mem_pf;
+    uint64_t io;
+    uint64_t mem_pf;
     uint64_t mem_pf_limit;
     uint8_t curr_bus_number;
 };
